Create MockRepositories and move methods from ActivateIntegrationWithCouponTest to it
